Dev zarnold nmea revamp (#12)
* add functionality for strings back into message decoder, with better explanation for what functions do

Remove reference to NMEA

* Fix ASCII_STRING_VALID Test

Initially, the start of a log was being ingested into our message decoder, this would never happen. As a result, the message decoder did not know how to handle the sync byte of the ASCII log, and would fail. I replaced the  test with string values of a different log's message portion.

* Add align boolean for byte alignment decision.  NMEA doesn't follow byte alignment